+++ src/game/Constants.h
@@ -1,7 +1,7 @@
 /*
  * Copyright 2009, 2010, The Pink Petal Development Team.
  * The Pink Petal Devloment Team are defined as the game's coders 
- * who meet on http://pinkpetal.org     // old site: http://pinkpetal .co.cc
+ * who meet on http://pinkpetal.org
  * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,7 +22,6 @@
 #include <array>
 #include <string>
 
-typedef unsigned int u_int;
 class sAttribute;
 
 enum {
@@ -76,7 +75,7 @@ enum FFSD        // `J` Free Fire Sell Dump
 
 // Editable Character Stats and skills (used for traits)
 //STATS enum
-enum STATS{
+enum STATS {
     STAT_CHARISMA        = 0    ,    //
     STAT_HAPPINESS            ,    //
     STAT_LIBIDO                ,    //